+++ src/pages/splash/Splash.tsx
